backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
openai>=1.30.0
qdrant-client>=1.11.0
pypdf==4.0.1
python-dotenv==1.0.1
langchain==0.1.6
langchain-openai==0.0.5
tiktoken==0.5.2
